Aggiornamento badge README e script per la compilazione dei badge
