+++ .github/workflows/main.yml
@@ -71,20 +71,20 @@ jobs:
       run: |
         cat .github/reports/pylint_report.txt
 
-    - name: Create Pylint Badge and put in README.md
+    - name: Create badges and put in README.md
       uses: actions/setup-python@v5
       with: 
         python-version: '3.12.2'
     - run: |
-        python pylintbadge.py
+        python readmeBadges.py
       
-    - name: Push README with changed badge to origin
+    - name: Push README with changed badges to origin
       run: |
         git config user.name "github-actions[bot]"
         git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
         if [[ $(git diff --name-only 'README.md') ]]; then
           git add README.md
-          git commit -m "Update Pylint Badge"
+          git commit -m "Update README badges"
           git push 
         else
           echo "No changes to README.md, skipping commit and push."
