+++ README.md
@@ -1,4 +1,5 @@
 # MVP
 [![Continuos Integration & Quality Checks](https://github.com/SevenBitsSwe/MVP/actions/workflows/main.yml/badge.svg)](https://github.com/SevenBitsSwe/MVP/actions/workflows/main.yml)
-[![Coverage Status](https://coveralls.io/repos/github/SevenBitsSwe/MVP/badge.svg)](https://coveralls.io/github/SevenBitsSwe/MVP)
 ![pylint](https://img.shields.io/badge/PyLint-8.96-green?logo=python&logoColor=white)
+![Line coverage](https://img.shields.io/badge/coverage-99.66%25-brightgreen?logo=python&logoColor=white)
+![Branch coverage](https://img.shields.io/badge/coverage-100.0%25-brightgreen?logo=python&logoColor=white)
